+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-ip-address-impl/src/main/java/com/liferay/multi/factor/authentication/ip/address/internal/checker/IPAddressHeadlessMFAChecker.java
@@ -5,11 +5,14 @@
 
 package com.liferay.multi.factor.authentication.ip.address.internal.checker;
 
-import com.liferay.multi.factor.authentication.ip.address.internal.audit.MFAIPAddressAuditMessageBuilder;
 import com.liferay.multi.factor.authentication.ip.address.internal.configuration.MFAIPAddressConfiguration;
+import com.liferay.multi.factor.authentication.ip.address.internal.constants.MFAIPAddressEventTypes;
 import com.liferay.multi.factor.authentication.spi.checker.headless.HeadlessMFAChecker;
-import com.liferay.osgi.util.service.Snapshot;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.audit.AuditException;
+import com.liferay.portal.kernel.audit.AuditMessage;
+import com.liferay.portal.kernel.audit.AuditRouterUtil;
+import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.User;
@@ -54,16 +57,11 @@ public class IPAddressHeadlessMFAChecker implements HeadlessMFAChecker {
 						userId);
 			}
 
-			MFAIPAddressAuditMessageBuilder mfaIPAddressAuditMessageBuilder =
-				_mfaIPAddressAuditMessageBuilderSnapshot.get();
-
-			if (mfaIPAddressAuditMessageBuilder != null) {
-				mfaIPAddressAuditMessageBuilder.routeAuditMessage(
-					mfaIPAddressAuditMessageBuilder.
-						buildNonexistentUserVerificationFailureAuditMessage(
-							CompanyThreadLocal.getCompanyId(), userId,
-							_getClassName()));
-			}
+			_mfaipAddressAuditMessageBuilder.routeAuditMessage(
+				_mfaipAddressAuditMessageBuilder.
+					buildNonexistentUserVerificationFailureAuditMessage(
+						CompanyThreadLocal.getCompanyId(), userId,
+						_getClassName()));
 
 			return false;
 		}
@@ -71,28 +69,18 @@ public class IPAddressHeadlessMFAChecker implements HeadlessMFAChecker {
 		if (AccessControlUtil.isAccessAllowed(
 				httpServletRequest, _allowedIpAddressesAndNetmasks)) {
 
-			MFAIPAddressAuditMessageBuilder mfaIPAddressAuditMessageBuilder =
-				_mfaIPAddressAuditMessageBuilderSnapshot.get();
-
-			if (mfaIPAddressAuditMessageBuilder != null) {
-				mfaIPAddressAuditMessageBuilder.routeAuditMessage(
-					mfaIPAddressAuditMessageBuilder.
-						buildVerificationSuccessAuditMessage(
-							user, _getClassName()));
-			}
+			_mfaipAddressAuditMessageBuilder.routeAuditMessage(
+				_mfaipAddressAuditMessageBuilder.
+					buildVerificationSuccessAuditMessage(
+						user, _getClassName()));
 
 			return true;
 		}
 
-		MFAIPAddressAuditMessageBuilder mfaIPAddressAuditMessageBuilder =
-			_mfaIPAddressAuditMessageBuilderSnapshot.get();
-
-		if (mfaIPAddressAuditMessageBuilder != null) {
-			mfaIPAddressAuditMessageBuilder.routeAuditMessage(
-				mfaIPAddressAuditMessageBuilder.
-					buildVerificationFailureAuditMessage(
-						user, _getClassName(), "IP is not allowed"));
-		}
+		_mfaipAddressAuditMessageBuilder.routeAuditMessage(
+			_mfaipAddressAuditMessageBuilder.
+				buildVerificationFailureAuditMessage(
+					user, _getClassName(), "IP is not allowed"));
 
 		return false;
 	}
@@ -136,15 +124,66 @@ public class IPAddressHeadlessMFAChecker implements HeadlessMFAChecker {
 	private static final Log _log = LogFactoryUtil.getLog(
 		IPAddressHeadlessMFAChecker.class);
 
-	private static final Snapshot<MFAIPAddressAuditMessageBuilder>
-		_mfaIPAddressAuditMessageBuilderSnapshot = new Snapshot<>(
-			IPAddressHeadlessMFAChecker.class,
-			MFAIPAddressAuditMessageBuilder.class);
-
 	private Set<String> _allowedIpAddressesAndNetmasks;
+	private final MFAIPAddressAuditMessageBuilder
+		_mfaipAddressAuditMessageBuilder =
+			new MFAIPAddressAuditMessageBuilder();
 	private ServiceRegistration<HeadlessMFAChecker> _serviceRegistration;
 
 	@Reference
 	private UserLocalService _userLocalService;
+
+	private class MFAIPAddressAuditMessageBuilder {
+
+		public AuditMessage buildNonexistentUserVerificationFailureAuditMessage(
+			long companyId, long userId, String mfaCheckerClassName) {
+
+			return new AuditMessage(
+				MFAIPAddressEventTypes.MFA_IP_ADDRESS_VERIFICATION_FAILURE,
+				companyId, userId, "Nonexistent", mfaCheckerClassName,
+				String.valueOf(userId), null,
+				JSONUtil.put("reason", "Nonexistent User"));
+		}
+
+		public AuditMessage buildVerificationFailureAuditMessage(
+			User user, String mfaCheckerClassName, String reason) {
+
+			return new AuditMessage(
+				MFAIPAddressEventTypes.MFA_IP_ADDRESS_VERIFICATION_FAILURE,
+				user.getCompanyId(), user.getUserId(), user.getFullName(),
+				mfaCheckerClassName, String.valueOf(user.getPrimaryKey()), null,
+				JSONUtil.put("reason", reason));
+		}
+
+		public AuditMessage buildVerificationSuccessAuditMessage(
+			User user, String mfaCheckerClassName) {
+
+			return new AuditMessage(
+				MFAIPAddressEventTypes.MFA_IP_ADDRESS_VERIFICATION_SUCCESS,
+				user.getCompanyId(), user.getUserId(), user.getFullName(),
+				mfaCheckerClassName, String.valueOf(user.getPrimaryKey()), null,
+				null);
+		}
+
+		public void routeAuditMessage(AuditMessage auditMessage) {
+			try {
+				AuditRouterUtil.route(auditMessage);
+			}
+			catch (AuditException auditException) {
+				if (_log.isWarnEnabled()) {
+					_log.warn("Unable to route audit message", auditException);
+				}
+			}
+			catch (Exception exception) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(exception);
+				}
+			}
+		}
+
+		private final Log _log = LogFactoryUtil.getLog(
+			MFAIPAddressAuditMessageBuilder.class);
+
+	}
 
 }
